+++ packages/authentication-local/test/strategy.test.ts
@@ -23,7 +23,7 @@ describe('@feathersjs/authentication-local/strategy', () => {
     try {
       auth.register('something', new LocalStrategy());
       assert.fail('Should never get here');
-    } catch (error) {
+    } catch (error: any) {
       assert.strictEqual(error.message,
         '\'something\' authentication strategy requires a \'usernameField\' setting'
       );
@@ -40,7 +40,7 @@ describe('@feathersjs/authentication-local/strategy', () => {
         password
       });
       assert.fail('Should never get here');
-    } catch (error) {
+    } catch (error: any) {
       assert.strictEqual(error.name, 'NotAuthenticated');
       assert.strictEqual(error.message, 'Invalid login');
     }
@@ -64,7 +64,7 @@ describe('@feathersjs/authentication-local/strategy', () => {
     try {
       await strategy.getEntity({}, {});
       assert.fail('Should never get here');
-    } catch (error) {
+    } catch (error: any) {
       assert.strictEqual(error.message, 'Could not get local entity');
     }
   });
@@ -91,7 +91,21 @@ describe('@feathersjs/authentication-local/strategy', () => {
         password: 'dummy'
       });
       assert.fail('Should never get here');
-    } catch (error) {
+    } catch (error: any) {
+      assert.strictEqual(error.name, 'NotAuthenticated');
+      assert.strictEqual(error.message, 'Invalid login');
+    }
+  });
+
+  it('fails when password is not provided', async () => {
+    const authService = app.service('authentication');
+    try {
+      await authService.create({
+        strategy: 'local',
+        email,
+      });
+      assert.fail('Should never get here');
+    } catch (error: any) {
       assert.strictEqual(error.name, 'NotAuthenticated');
       assert.strictEqual(error.message, 'Invalid login');
     }
@@ -111,7 +125,7 @@ describe('@feathersjs/authentication-local/strategy', () => {
         email: userEmail
       });
       assert.fail('Should never get here');
-    } catch (error) {
+    } catch (error: any) {
       assert.strictEqual(error.name, 'NotAuthenticated');
       assert.strictEqual(error.message, 'Invalid login');
     }