+++ packages/client/test/fetch.test.ts
@@ -1,3 +1,4 @@
+//@ts-ignore
 import fetch from 'node-fetch';
 import { Server } from 'http';
 import { clientTests } from '@feathersjs/tests';