+++ packages/rest-client/src/superagent.ts
@@ -23,7 +23,7 @@ export class SuperagentClient extends Base {
             const response = error.response;
             error = JSON.parse(error.response.text);
             error.response = response;
-          } catch (e) {}
+          } catch (e: any) {}
 
           return reject(error);
         }